requirements.txt
# Django==5.0.6
# mysqlclient==2.2.4
# python-dotenv==1.0.1
PyMuPDF
# faiss-cpu==1.12.0 
# numpy
# openai==1.40.3 
# tiktoken==0.7.0


# Python 3.13 compatible versions (Windows)

# Numpy stable version compatible with Py3.13
numpy

# MySQL client - use prebuilt wheel recommended
mysqlclient

# tiktoken - latest version compatible with Py3.13
tiktoken

# Faiss CPU version (skip 1.8.0.post1, use 1.12.0)
faiss-cpu

# Django stable version
Django

# Optional helpful packages
requests
python-dotenv
pandas